big q learning cleanup
